Simplify setGmcp and update tests
